+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+## 1.145.0 (October, 12 2023)
+ENHANCEMENTS:
+* resource/spotinst_ocean_aks_np: Added `pod_subnet_ids` and `vnet_subnet_ids` fields in `node_pool_properties` object
+* resource/spotinst_ocean_aks_np_virtual_node_group: Added `pod_subnet_ids` and `vnet_subnet_ids` fields in `node_pool_properties` object
+
 ## 1.144.0 (October, 09 2023)
 ENHANCEMENTS:
 * resource/spotinst_ocean_aws_launch_spec: Added `instancetypes_filters` object under `launchSpec`